drawAutoDetect/requirements.txt
turtle
opencv-python
numpy
Pillow
rembg
torch
transformers
scipy
realesrgan
basicsr 